# server/docs/swagger.yaml
definitions:
  consts.ConfigStatus:
    enum:
    - 0
    type: integer
    x-enum-comments:
      ConfigStatusEnable: 启用
    x-enum-varnames:
    - ConfigStatusEnable
...
  consts.DictStatus:
    enum:
    - 0
    type: integer
    x-enum-comments:
      DictStatusEnable: 启用
    x-enum-varnames:
    - DictStatusEnable
  consts.DictType:
    enum:
    - 0
    type: integer
    x-enum-comments:
      DictTypeInt: int
    x-enum-varnames:
    - DictTypeInt
...
  consts.UserStatus:
    enum:
    - 0
    type: integer
    x-enum-comments:
      UserStatusEnable: 启用
    x-enum-varnames:
    - UserStatusEnable
...
  dto.ResetPwdDto:
    properties:
      password:
        description: 密码
        type: string
      userId:
        description: 用户Id
        type: integer
    type: object
  dto.SetUserStateDto:
    properties:
      status:
        allOf:
        - $ref: '#/definitions/consts.UserStatus'
        description: 帐号状态（0正常 1停用）
      userId:
        description: 用户Id
        type: integer
    type: object
  dto.SysConfigAddDto:
    properties:
      configKey:
        description: 配置键名
        type: string
      configName:
        description: 配置名称
        type: string
      configValue:
        description: 配置键值
        type: string
      remark:
        description: 备注
        type: string
      status:
        allOf:
        - $ref: '#/definitions/consts.ConfigStatus'
        description: 配置状态（0正常 1停用）
    type: object
  dto.SysConfigUpdateDto:
    properties:
      configKey:
        description: 配置键名
        type: string
      configName:
        description: 配置名称
        type: string
      configValue:
        description: 配置键值
        type: string
      id:
        description: 编号
        type: integer
      remark:
        description: 备注
        type: string
      status:
        allOf:
        - $ref: '#/definitions/consts.ConfigStatus'
        description: 配置状态（0正常 1停用）
    type: object
  dto.SysConfigVo:
    properties:
      configKey:
        description: 配置键名
        type: string
      configName:
        description: 配置名称
        type: string
      configValue:
        description: 配置键值
        type: string
      createdAt:
        description: 创建时间
        type: string
      id:
        description: 配置id
        type: integer
      remark:
        description: 备注
        type: string
      status:
        allOf:
        - $ref: '#/definitions/consts.ConfigStatus'
        description: 配置状态（0正常 1停用）
      updatedAt:
        description: 更新时间
        type: string
    type: object
  dto.SysDictAddDto:
    properties:
      dictCode:
        description: 字典代码
        type: string
      dictName:
        description: 字典名称
        type: string
      dictType:
        allOf:
        - $ref: '#/definitions/consts.DictType'
        description: 字典类型（0int，1string）
      remark:
        description: 备注
        type: string
      status:
        allOf:
        - $ref: '#/definitions/consts.DictStatus'
        description: 字典状态（0正常 1停用）
    type: object
...
  dto.SysDictUpdateDto:
    properties:
      dictCode:
        description: 字典代码
        type: string
      dictName:
        description: 字典名称
        type: string
      dictType:
        allOf:
        - $ref: '#/definitions/consts.DictType'
        description: 字典类型（0int，1string）
      id:
        description: 编号
        type: integer
      remark:
        description: 备注
        type: string
      status:
        allOf:
        - $ref: '#/definitions/consts.DictStatus'
        description: 字典状态（0正常 1停用）
    type: object
  dto.SysDictVo:
    properties:
      createdAt:
        description: 创建时间
        type: string
      dictCode:
        description: 字典代码
        type: string
      dictName:
        description: 字典名称
        type: string
      dictType:
        allOf:
        - $ref: '#/definitions/consts.DictType'
        description: 字典类型（0int，1string）
      id:
        description: 编号
        type: integer
      remark:
        description: 备注
        type: string
      status:
        allOf:
        - $ref: '#/definitions/consts.DictStatus'
        description: 字典状态（0正常 1停用）
      updatedAt:
        description: 更新时间
        type: string
    type: object
...
  dto.SysUserAddDto:
    properties:
      email:
        description: 用户邮箱
        type: string
      nickName:
        description: 用户昵称
        type: string
      password:
        description: 密码
        type: string
      phone:
        description: 手机号码
        type: string
      remark:
        description: 备注
        type: string
      roleIds:
        description: 角色
        items:
          type: integer
        type: array
      sex:
        description: 用户性别（0未知，1男，2女）
        type: integer
      status:
        allOf:
        - $ref: '#/definitions/consts.UserStatus'
        description: 帐号状态（0正常 1停用）
      userName:
        description: 用户账号
        type: string
    type: object
  dto.SysUserUpdateDto:
    properties:
      email:
        description: 用户邮箱
        type: string
      id:
        description: 编号
        type: integer
      nickName:
        description: 用户昵称
        type: string
      phone:
        description: 手机号码
        type: string
      remark:
        description: 备注
        type: string
      roleIds:
        description: 角色
        items:
          type: integer
        type: array
      sex:
        description: 用户性别（0未知，1男，2女）
        type: integer
      status:
        allOf:
        - $ref: '#/definitions/consts.UserStatus'
        description: 帐号状态（0正常 1停用）
    type: object
  dto.SysUserVo:
    properties:
      avatar:
        description: 头像地址
        type: string
      createdAt:
        description: 创建时间
        type: string
      email:
        description: 用户邮箱
        type: string
      id:
        description: 编号
        type: integer
      nickName:
        description: 用户昵称
        type: string
      phone:
        description: 手机号码
        type: string
      remark:
        description: 备注
        type: string
      roleIds:
        description: 角色
        items:
          type: integer
        type: array
      sex:
        description: 用户性别（0未知，1男，2女）
        type: integer
      status:
        description: 帐号状态（0正常 1停用）
        type: integer
      updatedAt:
        description: 更新时间
        type: string
      userName:
        description: 用户账号
        type: string
      userType:
        description: 用户类型（0普通账号，1超级管理员）
        type: integer
    type: object
...
  request.IdInfoDto:
    properties:
      id:
        description: id编号
        type: integer
    type: object
  response.JsonResult:
    properties:
      code:
        description: 状态码（架构层面：中间件/异常处理等）
        type: integer
      data:
        description: 数据
      message:
        description: 消息
        type: string
      success:
        description: 是否成功（业务层面）
        type: boolean
    type: object
  response.PageResult:
    properties:
      list:
        description: 数据
      totalCount:
        description: 总条数
        type: integer
    type: object
info:
  contact: {}
  description: go-fast-admin 接口文档
  title: 系统模块
  version: 1.0.0
paths:
...
  /system/config/add:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.SysConfigAddDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 添加配置
      tags:
      - 配置
  /system/config/delete:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdInfoDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 删除配置
      tags:
      - 配置
  /system/config/detail:
    get:
      parameters:
      - description: id编号
        in: query
        name: id
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  $ref: '#/definitions/dto.SysConfigVo'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取配置详情
      tags:
      - 配置
  /system/config/list:
    get:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.SysConfigVo'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: 配置列表
      tags:
      - 配置
  /system/config/query:
    get:
      parameters:
      - description: 配置名称
        in: query
        name: configName
        type: string
      - description: 页码
        in: query
        name: pageIndex
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/dto.SysConfigVo'
                        type: array
                    type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: 配置查询
      tags:
      - 配置
  /system/config/update:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.SysConfigUpdateDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 更新配置
      tags:
      - 配置
  /system/dict/add:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.SysDictAddDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 添加字典
      tags:
      - 字典
  /system/dict/delete:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdInfoDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 删除字典
      tags:
      - 字典
  /system/dict/detail:
    get:
      parameters:
      - description: id编号
        in: query
        name: id
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  $ref: '#/definitions/dto.SysDictVo'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取字典详情
      tags:
      - 字典
  /system/dict/list:
    get:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.SysDictVo'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: 字典列表
      tags:
      - 字典
  /system/dict/query:
    get:
      parameters:
      - description: 字典名称
        in: query
        name: dictName
        type: string
      - description: 页码
        in: query
        name: pageIndex
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/dto.SysDictVo'
                        type: array
                    type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: 字典查询
      tags:
      - 字典
  /system/dict/update:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.SysDictUpdateDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 更新字典
      tags:
      - 字典
...
  /system/user/add:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.SysUserAddDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 添加用户
      tags:
      - 用户
  /system/user/delete:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdInfoDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 删除用户
      tags:
      - 用户
  /system/user/detail:
    get:
      parameters:
      - description: id编号
        in: query
        name: id
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  $ref: '#/definitions/dto.SysUserVo'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取用户详情
      tags:
      - 用户
  /system/user/list:
    get:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.SysUserVo'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: 用户列表
      tags:
      - 用户
  /system/user/query:
    get:
      parameters:
      - description: 用户昵称
        in: query
        name: nickName
        type: string
      - description: 页码
        in: query
        name: pageIndex
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/dto.SysUserVo'
                        type: array
                    type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: 用户查询
      tags:
      - 用户
  /system/user/resetPwd:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPwdDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 重置密码
      tags:
      - 用户
  /system/user/setStatus:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.SetUserStateDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 设置用户状态
      tags:
      - 用户
  /system/user/update:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.SysUserUpdateDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 更新用户
      tags:
      - 用户
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"